Fixed the issue with multimember setup for test cases switching members
