+++ Hazelcast.Test/Hazelcast.Client.Test/ClientPNCounterConsistencyLossTest.cs
@@ -1,4 +1,18 @@
-﻿using System.Linq;
+﻿// Copyright (c) 2008-2018, Hazelcast, Inc. All Rights Reserved.
+// 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+// http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System.Linq;
 using System.Threading;
 using Hazelcast.Client.Proxy;
 using Hazelcast.Config;
@@ -10,22 +24,27 @@ namespace Hazelcast.Client.Test
 {
     [TestFixture("AddAndGet")]
     [TestFixture("Get")]
-    [TestFixture("DriverCanContinueSessionByCallingReset")]
-    public class ClientPNCounterConsistencyLossTest : MultiMemberBaseTest
+    public class ClientPNCounterConsistencyLossTest : MultiMemberBaseNoSetupTest
     {
         private readonly string _type;
+        private ClientPNCounterProxy _pnCounter;
 
         public ClientPNCounterConsistencyLossTest(string type)
         {
             _type = type;
         }
 
-        private ClientPNCounterProxy _pnCounter;
-
         [SetUp]
         public void Setup()
         {
+            SetupCluster();
             _pnCounter = Client.GetPNCounter(TestSupport.RandomString()) as ClientPNCounterProxy;
+        }
+
+        [TearDown]
+        public void TearDown()
+        {
+            ShutdownRemoteController();
         }
 
         protected override void InitMembers()
@@ -47,6 +66,19 @@ namespace Hazelcast.Client.Test
         }
 
         [Test]
+        public void DriverCanContinueSessionByCallingReset()
+        {
+            _pnCounter.AddAndGet(5);
+            Assert.AreEqual(5, _pnCounter.Get());
+
+            TerminateTargetReplicaMember();
+            Thread.Sleep(1000);
+
+            _pnCounter.Reset();
+            Mutation();
+        }
+
+        [Test]
         public void ConsistencyLostExceptionIsThrownWhenTargetReplicaDisappears()
         {
             _pnCounter.AddAndGet(5);
@@ -55,7 +87,7 @@ namespace Hazelcast.Client.Test
             TerminateTargetReplicaMember();
             Thread.Sleep(1000);
 
-            Mutation();
+            Assert.Throws<ConsistencyLostException>(Mutation);
         }
 
         private void Mutation()
@@ -63,16 +95,11 @@ namespace Hazelcast.Client.Test
             switch (_type)
             {
                 case "AddAndGet":
-                    Assert.Throws<ConsistencyLostException>(() => _pnCounter.AddAndGet(5));
+                    _pnCounter.AddAndGet(5);
                     break;
 
                 case "Get":
-                    Assert.Throws<ConsistencyLostException>(() => _pnCounter.Get());
-                    break;
-
-                case "DriverCanContinueSessionByCallingReset":
-                    _pnCounter.Reset();
-                    _pnCounter.AddAndGet(5);
+                    _pnCounter.Get();
                     break;
             }
         }
